requirements.txt
# Colloq — Python dependencies
# Install with: pip install -r requirements.txt

streamlit>=1.33.0       # The web framework that renders all UI components
openai>=1.30.0          # OpenAI-compatible SDK — used to call OpenRouter's API
python-dotenv>=1.0.0    # Loads environment variables from a .env file
streamlit-float>=0.3.0  # Floating containers / pinned UI elements
